add item, fixed add user

+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -41,20 +41,18 @@
 		<tbody>
 			<thead>
 				<tr>
-					<th>Item</th><th>Description</th><th>Quantity</th><th>Action</th>
+					<th>Item</th><th>Description</th><th>Quantity</th><th>Price</th><th>Action</th>
 				</tr>
 			</thead>
 			
 				<c:forEach var="item" items="${items}">
-				<tr>
-					<%-- <td><a href="/items/${item.id}">${item.name}</a></td>--%>
-							
+				<tr>							
 					
-					<td>${item.name}</td><td>${item.description}</td><td>${item.quantity}</td>
+					<td>${item.name}</td><td>${item.description}</td><td>${item.quantity}</td><td>$ ${item.price}</td>
 					<td>
 								 <%--  // added/items to see if working --%>
 						<a href="/items/${ item.id }/edit" class="btn btn-light btn-sm">Edit</a>
-						<a href="/items/${ item.id }/delete" class="btn btn-light btn-sm">Delete</a>
+						<a onclick="return confirm('Are you sure you want to delete?')" href="/items/${ item.id }/delete" class="btn btn-light btn-sm">Delete</a>
 					</td>
 				</tr>
 				</c:forEach>
